Refactoring to remove code redundancy
