+++ src/pages/datamodel/ImageMetadata/ImageMetadata.tsx
@@ -1,8 +1,6 @@
-import { PropertyDefinitionEditorDialog, moveArrayItem } from '@/components/PropertyDefinitionEditor';
 import { PropertyDefinition } from '@/model';
 import { useDataModel } from '@/store';
-import { Button } from '@/ui/Button';
-import { MetadataListItem } from '../MetadataListItem';
+import { Metadata } from '../Metadata';
 
 export const ImageMetadata = () => {
 
@@ -30,53 +28,12 @@ export const ImageMetadata = () => {
       });
     }
   }
-
-  const addProperty = (added: PropertyDefinition) =>
-    onChange([...properties, added]);
-
-  const moveProperty = (definition: PropertyDefinition, up: boolean) =>
-    onChange(moveArrayItem(properties, properties.indexOf(definition), up));
-
-  const updateProperty = (updated: PropertyDefinition, previous: PropertyDefinition) =>
-    onChange(properties.map(p => p === previous ? updated : p));
-
-  const deleteProperty = (definition: PropertyDefinition) =>
-    onChange(properties.filter(d => d !== definition));
-
   return (
-    <>
-      <p className="p-1 mt-4 text-sm max-w-xl leading-6">
-        Define a metadata schema to record structured information about your images. 
-      </p>
-
-      {properties.length === 0 ? (
-        <div className="h-8 mt-3" />
-      ) : (
-        <ul className="max-w-sm mt-8 mb-12">
-          {properties.map(definition => (
-            <li key={definition.name}>
-              <MetadataListItem 
-                editorHint={editorHint}
-                previewHint={previewHint}
-                definition={definition} 
-                onMoveUp={() => moveProperty(definition, true)}
-                onMoveDown={() => moveProperty(definition, false)}
-                onUpdateProperty={updated => updateProperty(updated, definition)}
-                onDeleteProperty={() => deleteProperty(definition)} />
-            </li>
-          ))}
-        </ul>
-      )}
-
-      <PropertyDefinitionEditorDialog
-        editorHint={editorHint}
-        previewHint={previewHint}
-        onSave={addProperty}>
-        <Button>
-          Add Image Property
-        </Button>
-      </PropertyDefinitionEditorDialog>
-    </>
+    <Metadata 
+      editorHint={editorHint} 
+      previewHint={previewHint} 
+      properties={properties} 
+      onChange={onChange} />
   )
 
 }
